requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
numpy
scikit-learn
google-genai
markdown-it-py
pytest
